pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bbs"
version = "0.0.1"
dependencies = [
  "click",
  "litestar",
  "piccolo[sqlite]",
  "aiosqlite",
  "pynacl",
  "pyjwt",
  "python-jose",
  "sqlmodel",
  "starlette",
  "toml",
  "uvicorn",
  "textual",
  "httpx",
  "websockets",
  "python-dotenv",
  "rich",
  "rich-click",
  "markdown-it-py",
  "litestar-htmx",
  "anyio",
]
requires-python = ">=3.8"
authors = [
  {name = "merx"},
]
# maintainers = [
#   {name = "Brett Cannon", email = "brett@example.com"}
# ]
description = "bbs system"
# readme = "README.rst"
readme = "README.md"
# license = {file = "LICENSE.txt"}
# keywords = ["egg", "bacon", "sausage", "tomatoes", "Lobster Thermidor"]
classifiers = [
  # "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]

# [project.optional-dependencies]
# gui = ["PyQt5"]
# cli = [
#   "rich",
#   "click",
# ]

# [project.urls]
# Homepage = "https://example.com"
# Documentation = "https://readthedocs.org"
# Repository = "https://github.com/me/spam.git"
# "Bug Tracker" = "https://github.com/me/spam/issues"
# Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[project.scripts]
bbs-cli = "bbs:cli"
bbs-tui = "bbs_client.app:run"
# bbs-cli = "bbs.__main__:cli"
# bbs-cli = "bbs:run_uvicorn"
# spam-cli = "spam:main_cli"

# [project.gui-scripts]
# spam-gui = "spam:main_gui"

# [project.entry-points."spam.magical"]
# tomatoes = "spam:main_tomatoes"

[tool.pyright]
# reportIncompatibleVariableOverride = "information"
reportIncompatibleVariableOverride = "none"
